requirements.txt
pinecone
google-generativeai
together
PyPDF2
fpdf
duckduckgo-search
openai
python-dotenv
chardet
markdown